+++ src/app/inobeta-ui/ui/material-forms/controls/dropdown.spec.ts
@@ -2,7 +2,7 @@ import { async, TestBed } from '@angular/core/testing';
 
 import { IbToolTestModule } from '../../../tools';
 import { CommonModule } from '@angular/common';
-import { FormControl, FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormBuilder, FormControl, FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { IbDynamicFormsModule } from '../../forms';
 import { MatFormFieldModule, MatOptionModule, MatSelectModule } from '@angular/material';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
@@ -13,6 +13,9 @@ import {  IbMatDropdownComponent } from '../controls/dropdown';
 
 describe('IbMatDropdownComponent', () => {
   let component: IbMatDropdownComponent;
+  let formBuilder: FormBuilder;
+  let control: FormControl;
+  let form: FormGroup;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -33,13 +36,16 @@ describe('IbMatDropdownComponent', () => {
       ]
     })
     .compileComponents();
+    formBuilder = TestBed.get(FormBuilder);
+    control = new FormControl('control');
+    form = formBuilder.group(control);
   }));
 
   beforeEach(() => {
     const componentRef = TestBed.createComponent(IbMatDropdownComponent);
     component = componentRef.componentInstance;
     component.data = {
-      self: new FormControl(),
+      self: control,
       base: {
         options: [
           { key: 'test1', value: 'value1' },
@@ -48,10 +54,10 @@ describe('IbMatDropdownComponent', () => {
         multiple: true,
         change: () => {}
       },
-      form: null,
+      form,
       hasError: null,
       formControlErrors: null
-    }
+    };
   });
 
   it('should create', () => {
@@ -61,11 +67,11 @@ describe('IbMatDropdownComponent', () => {
 
   it('should select and deselect all', () => {
     const spy = spyOn(component.data.self, 'setValue').and.callThrough();
-    component.selectAll()
-    expect(spy).toHaveBeenCalledWith(['test1', 'test2'])
+    component.selectAll();
+    expect(spy).toHaveBeenCalledWith(['test1', 'test2']);
     spy.calls.reset();
-    component.selectAll()
-    expect(spy).toHaveBeenCalledWith([])
+    component.selectAll();
+    expect(spy).toHaveBeenCalledWith([]);
     spy.calls.reset();
   });
 
@@ -73,16 +79,18 @@ describe('IbMatDropdownComponent', () => {
 
   it('should handle selection when multiple', () => {
     const spy = spyOn(component.data.base, 'change').and.callThrough();
-    component.handleSelection({value: []})
-    expect(spy).not.toHaveBeenCalled()
+    component.handleSelection({value: []});
+    expect(spy).not.toHaveBeenCalled();
     spy.calls.reset();
-    component.handleSelection({value: ['something']})
-    expect(spy).toHaveBeenCalled()
+    component.handleSelection({value: ['something']});
+    expect(spy).toHaveBeenCalled();
     spy.calls.reset();
     component.data.base.multiple = false;
-    component.handleSelection({value: []})
-    expect(spy).toHaveBeenCalled()
+    component.handleSelection({value: []});
+    expect(spy).toHaveBeenCalled();
     spy.calls.reset();
+    control.setValue(['test1', 'test2']);
+    component.handleSelection({value: ['test1', 'test2']});
   });
 
 });