+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Inobeta/UI Changelog
 
+## [1.1.6]
+
+- fix production build on wrong toast import
+- IbDynamicForms: debounceOnChange default to 0
+- IbHttpModule: added ibHttpUrlExcludedFromLoader spinner loading url exclusions
+- IbMatTextboxComponent get min/max DOM validation from angular validators
+- IbMatSlideToggleComponent added
+- auth interceptor now force logout on 401 errors
+
 ## [1.1.5]
 
 - numbers type added to IbMatDropdownControl